Place execute callback inside the setup callback so that intermediate passdata is no longer required

+++ D3D12/Graphics/Texture.cpp
@@ -422,7 +422,7 @@ void Texture::Create(CommandContext* pContext, const char* pFilePath, bool srgb)
 			pImg = pImg->GetNextImage();
 		}
 		Create(desc);
-		pContext->InitializeTexture(this, subResourceData.data(), 0, subResourceData.size());
+		pContext->InitializeTexture(this, subResourceData.data(), 0, (int)subResourceData.size());
 		pContext->ExecuteAndReset(true);
 	}
 }
